requirements.txt
click
plotly
python-snappy
zstd
